Feature/board

+++ b2b/src/main/java/kh/gangnam/b2b/service/ServiceImpl/BoardServiceImpl.java
@@ -1,24 +1,23 @@
 package kh.gangnam.b2b.service.ServiceImpl;
 
-import jakarta.persistence.EntityNotFoundException;
 import kh.gangnam.b2b.dto.board.request.*;
 import kh.gangnam.b2b.dto.board.response.CommentSaveResponse;
 import kh.gangnam.b2b.dto.board.response.CommentUpdateResponse;
 import kh.gangnam.b2b.dto.board.response.EditResponse;
-import kh.gangnam.b2b.dto.board.response.MessageResponse;
+import kh.gangnam.b2b.dto.MessageResponse;
 import kh.gangnam.b2b.dto.s3.S3Response;
 import kh.gangnam.b2b.entity.auth.Employee;
+import kh.gangnam.b2b.exception.NotFoundException;
 import kh.gangnam.b2b.repository.board.*;
 import kh.gangnam.b2b.service.BoardService;
 import kh.gangnam.b2b.service.shared.EmployeeCommonService;
 import kh.gangnam.b2b.util.S3ServiceUtil;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.PageRequest;
 import kh.gangnam.b2b.entity.board.*;
 import java.util.List;
 
@@ -40,6 +39,7 @@ public class BoardServiceImpl implements BoardService {
         // 작성 employee 가져오기
         Employee employee = employeeCommonService
                 .getEmployeeOrThrow(employeeId, "해당 사원을 찾을 수 없습니다.");
+
         // DB에 게시물 저장
         Board board = boardRepo.save(saveRequest.toEntity(employee));
         String content = saveRequest.content();
@@ -76,8 +76,13 @@ public class BoardServiceImpl implements BoardService {
         // id로 해당 employee,board,comment 찾기
         Employee employee = employeeCommonService
                 .getEmployeeOrThrow(employeeId, "해당 사원을 찾을 수 없습니다.");
-        Board board = boardRepo.findById(dto.boardId()).orElseThrow();
-        Comment parent = (dto.parentId() != null)? commentRepo.findById(dto.parentId()).orElseThrow():null;
+
+        Board board = (dto.boardId() != null)?boardRepo
+                .findById(dto.boardId()).orElseThrow(() -> new NotFoundException("게시판을 찾을 수 없습니다")):null;
+
+        Comment parent = (dto.parentId() != null)?commentRepo
+                .findById(dto.parentId()).orElseThrow(() -> new NotFoundException("댓글을 찾을 수 없습니다")):null;
+
 
         // comment 테이블에 전달된 정보 저장
         Comment comment = dto.toEntity(board,employee,parent);
@@ -90,7 +95,8 @@ public class BoardServiceImpl implements BoardService {
     public List<CommentSaveResponse> getCommentList(Long boardId, Long employeeId) {
 
         // 보드에서 댓글 List로 불러오기
-        List<Comment> comment = boardRepo.findById(boardId).orElseThrow().getComments();
+        List<Comment> comment = boardRepo
+                .findById(boardId).orElseThrow(() -> new NotFoundException("게시판을 찾을 수 없습니다")).getComments();
 
         // dto로 변환해서 리턴
         return comment.stream().map((commentSave)->{
@@ -101,6 +107,9 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public MessageResponse commentDeleteBoard(Long commentId) {
 
+        if (!commentRepo.existsById(commentId)) {
+            throw new NotFoundException("삭제할 댓글이 존재하지 않습니다");
+        }
         // 해당 댓글 삭제
         commentRepo.deleteById(commentId);
 
@@ -111,9 +120,9 @@ public class BoardServiceImpl implements BoardService {
     @Transactional
     public CommentUpdateResponse updateComment(CommentUpdateRequest dto, Long employeeId) {
 
-        System.out.println(dto);
         // 해당하는 댓글 entity 찾기
-        Comment comment = commentRepo.findById(dto.commentId()).orElseThrow();
+        Comment comment = commentRepo.findById(dto.commentId())
+                .orElseThrow(()-> new NotFoundException("댓글이 존재하지 않습니다"));
 
         comment.setComment(dto.comment());
 
@@ -121,24 +130,28 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public List<BoardResponse> getListBoard(int type, int page) {
+    public List<CommentSaveResponse> getReplyList(Long commentId, Long employeeId) {
+
+        return commentRepo.findById(commentId).orElseThrow(()-> new NotFoundException("댓글이 존재하지 않습니다"))
+                .getChildren().stream().map((comment)->CommentSaveResponse.fromEntity(comment,employeeId)).toList();
+    }
+
+    @Override
+    public Page<BoardResponse> getListBoard(int type, Pageable pageable) {
         BoardType boardType = BoardType.useTypeNo(type);
 
-        Sort sort=Sort.by(Sort.Direction.DESC, "boardId");
-        Pageable pageable= PageRequest.of(page-1,10, sort);
-
-        return boardRepo.findAllByType(boardType,pageable).stream()
-                .map(BoardResponse::fromEntity)
-                .toList();
+        return boardRepo.findByType(boardType, pageable)
+                .map(BoardResponse::fromEntity);
     }
 
     @Override
     public BoardResponse getBoard(Long boardId,Long employeeId) {
+
         return boardRepo.findById(boardId)
                 .map((board)->{
                     return BoardResponse.fromEntity(board,employeeId);
                 })
-                .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 게시글입니다."));
+                .orElseThrow(()-> new NotFoundException("게시글이 존재하지 않습니다"));
     }
 
     @Transactional
@@ -146,7 +159,9 @@ public class BoardServiceImpl implements BoardService {
     public BoardResponse updateBoard(Long boardId, UpdateRequest request) {
 
         // 게시글 정보 조회 후 저장
-        Board board = boardRepo.findById(boardId).orElseThrow().update(request);
+        Board board = boardRepo.findById(boardId)
+                .orElseThrow(() -> new NotFoundException("게시판을 찾을 수 없습니다")).update(request);
+
         String content = request.content();
 
         s3ServiceUtil.deleteBoardImage(boardId);
@@ -180,6 +195,10 @@ public class BoardServiceImpl implements BoardService {
     @Transactional
     public MessageResponse deleteBoard(Long boardId) {
 
+        if (!boardRepo.existsById(boardId)) {
+            throw new NotFoundException("삭제할 게시글이 존재하지 않습니다");
+        }
+
         // 게시물과 s3에 업로드된 이미지 삭제
         boardRepo.deleteById(boardId);
         s3ServiceUtil.deleteBoardImage(boardId);
@@ -191,8 +210,9 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public EditResponse editBoard(Long boardId) {
 
-        Board board = boardRepo.findById(boardId).orElseThrow(()
-                -> new RuntimeException("해당 게시글을 찾을 수 없습니다."));
+        Board board = boardRepo.findById(boardId)
+                .orElseThrow(() -> new NotFoundException("게시판을 찾을 수 없습니다"));
+
         return s3ServiceUtil.editBoardUrl(board);
     }
